BUG/MINOR: ssl: support tune.ssl.cachesize 0 again
Since the split of the shctx and the ssl cache, we lost the ability to
disable the cache with tune.ssl.cachesize 0.

Worst than that, when using this configuration, haproxy segfaults during
the configuration parsing.

Must be backported to 1.8.
(cherry picked from commit c61c0b371bebe0619cf4628b60b7c79d19e97269)

Signed-off-by: Willy Tarreau <w@1wt.eu>