BUG/MEDIUM: threads/vars: Fix deadlock in register_name
In register_name, before locking the var_names array, we check the variable name
validity. So if we try to register an invalid or empty name, we need to return
without unlocking it (because it was never locked).

This patch must be backported in 1.8.
(cherry picked from commit eb3e276d3911c4c95d1e2f270bab354440378701)

Signed-off-by: Willy Tarreau <w@1wt.eu>