MINOR: h2: store the demux padding length in the h2c struct
We'll try to process partial frames and for this we need to know the
padding length. The first step requires to extract it during the parsing
and store it in the demux context in the connection. Till now it was only
processed at once.

(cherry picked from commit 05e5dafe9a3fb2c07fc9bcb1e5c684ce2d7c4f33)
Signed-off-by: Willy Tarreau <w@1wt.eu>