BUG/MINOR: stream-int: don't try to receive again after receiving an EOS
When an end of stream has been reported, we should not try to receive again
as the mux layer might not be prepared to this and could report unexpected
errors.

This is more of a strengthening measure that follows the introduction of
conn_stream that came in 1.8. It's desired to backport this into 1.8 though
it's uncertain at this time whether it may have caused real issues.

(cherry picked from commit b78b80efe57fb7c6e195850626065d55ed5e4e3b)
Signed-off-by: Willy Tarreau <w@1wt.eu>