BUG/MINOR: http: properly detect max-age=0 and s-maxage=0 in responses
In 1.3.8, commit a15645d ("[MAJOR] completed the HTTP response processing.")
improved the response parser by taking care of the cache-control header
field. The parser is wrong because it is split in two parts, one checking
for elements containing an equal sign and the other one for those without.
The "max-age=0" and "s-maxage=0" tests were located at the wrong place and
thus have never matched. In practice the side effect was very minimal given
that this code used to be enabled only when checking if a cookie had the
risk of being cached or not. Recently in 1.8 it was also used to decide if
the response could be cached but in practice the cache takes care of these
values by itself so there is very limited impact.

This fix can be backported to all stable versions.
(cherry picked from commit d3900cc31d4ca14a9c45902b2543da2f39ce2c26)

Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit b396103341dda1b73ff43f84ea8cdf270710db6f)

Signed-off-by: Willy Tarreau <w@1wt.eu>