MINOR: dns: Handle SRV record weight correctly.
A SRV record weight can range from 0 to 65535, while haproxy weight goes
from 0 to 256, so we have to divide it by 256 before handing it to haproxy.
Also, a SRV record with a weight of 0 doesn't mean the server shouldn't be
used, so use a minimum weight of 1.

This should probably be backported to 1.8.

(cherry picked from commit 2ec2db9725fb54a76c726c5b8cc502071c575d28)
Signed-off-by: Willy Tarreau <w@1wt.eu>